Bonferroni and Simultaneous Intervals
The dataset is of Milk Transportation Cost. There are 3 variables,
Fuel, Repair and Capital. We first construct QQ plots of the variables
to determine if the data looks normal. Outliers are detected and
removed from the dataset. We construct new QQ plots without the
outliers and analyze the plots again. We also look at pairwise
scatterplots of the variables. At the end, Bonferroni and Simultaneous
Intervals are constructed for the data with outliers and a new dataset
without outliers. Comparisons are made.